fix(selene): Fixed a regression introduced in v1.1.2
The NFT that caused the previous error is still indexed wrong.
Instead of being assigned to the '8yp...' it's get assigned to the
minter/creator wallet which is the only one available from the APIs.

The current version doen't update ownership data with placeholder if
not available, this fixes the problem but requires manual rollback
to the previously correct version.